+++ fathom-archetype-standard/src/main/resources/archetype-resources/src/test/java/conf/RoutesTest.java
@@ -16,7 +16,7 @@
 
 package ${package}.conf;
 
-import fathom.test.FathomTest;
+import fathom.test.FathomIntegrationTest;
 import org.junit.Test;
 
 import static com.jayway.restassured.RestAssured.*;
@@ -26,7 +26,7 @@ import static org.hamcrest.Matchers.*;
  * Tests that our index page is generated.
  * Each unit test starts an instance of our Fathom app in TEST mode.
  */
-public class RoutesTest extends FathomTest {
+public class RoutesTest extends FathomIntegrationTest {
 
     @Test
     public void testIndex() {