+++ fathom-integration-test/src/test/java/conf/RoutesTest.java
@@ -16,12 +16,12 @@
 
 package conf;
 
-import fathom.test.FathomTest;
+import fathom.test.FathomIntegrationTest;
 import org.junit.Test;
 
 import static com.jayway.restassured.RestAssured.get;
 
-public class RoutesTest extends FathomTest {
+public class RoutesTest extends FathomIntegrationTest {
 
     @Test
     public void testIndex() {